Probando la verificación de rol

+++ backend/app/main/views.py
@@ -7,79 +7,90 @@ main_bp = Blueprint('main', __name__)
 
 @main_bp.route('/api/crearConductor', methods=['POST'])
 @verificar_token
-def crear_conductor():
-    from models import Conductor
+def crear_conductor(id_rol):
+    if(id_rol == 2):
+        from models import Conductor
 
-    # Acceder a los campos del formulario
-    nombre = request.form.get('nombre')
-    patente = request.form.get('patente')
-    auto = request.form.get('auto')
+        # Acceder a los campos del formulario
+        nombre = request.form.get('nombre')
+        patente = request.form.get('patente')
+        auto = request.form.get('auto')
 
-    # Verificar si la patente ya está en uso
-    if Conductor.query.filter_by(patente=patente).first():
-        return jsonify({'error': 'La patente ya está registrada'}), 400
+        # Verificar si la patente ya está en uso
+        if Conductor.query.filter_by(patente=patente).first():
+            return jsonify({'error': 'La patente ya está registrada'}), 400
 
-    # Crear un nuevo conductor si la patente no está en uso
-    nuevo_conductor = Conductor(nombre_conductor=nombre,patente=patente, nombre_vehiculo=auto)
-    db.session.add(nuevo_conductor)
-    db.session.commit()
+        # Crear un nuevo conductor si la patente no está en uso
+        nuevo_conductor = Conductor(nombre_conductor=nombre,patente=patente, nombre_vehiculo=auto)
+        db.session.add(nuevo_conductor)
+        db.session.commit()
 
-    # Guardar la foto del conductor si se adjuntó
-    foto_conductor = request.files.get('foto')
-    if foto_conductor:
-        nuevo_conductor.save_foto(foto_conductor)
+        # Guardar la foto del conductor si se adjuntó
+        foto_conductor = request.files.get('foto')
+        if foto_conductor:
+            nuevo_conductor.save_foto(foto_conductor)
 
-    return jsonify({'mensaje': 'Conductor creado correctamente'}), 201
+        return jsonify({'mensaje': 'Conductor creado correctamente'}), 201
+    else:
+        return jsonify({'mensaje': 'No tienes el permiso para crear un conductor'}), 400
 
 @main_bp.route('/api/obtenerConductores', methods=['GET'])
 @verificar_token
-def obtener_conductores():
-    from models import Conductor
-    conductores = Conductor.query.all()
-    conductores_data = []
-    for conductor in conductores:
-        foto_url = request.url_root + conductor.foto if conductor.foto else None
-        conductor_data = {
-            'id': conductor.id,
-            'nombre': conductor.nombre_conductor,
-            'patente': conductor.patente,
-            'auto': conductor.nombre_vehiculo,
-            'foto': foto_url
-        }
-        conductores_data.append(conductor_data)
-    return jsonify({'conductor': conductores_data})
+def obtener_conductores(id_rol):
+    if(id_rol == 2):
+        from models import Conductor
+        conductores = Conductor.query.all()
+        conductores_data = []
+        for conductor in conductores:
+            foto_url = request.url_root + conductor.foto if conductor.foto else None
+            conductor_data = {
+                'id': conductor.id,
+                'nombre': conductor.nombre_conductor,
+                'patente': conductor.patente,
+                'auto': conductor.nombre_vehiculo,
+                'foto': foto_url
+            }
+            conductores_data.append(conductor_data)
+        return jsonify({'conductor': conductores_data})
+    else:
+        return jsonify({'error': 'No tienes el permiso para obtener los conductores'})
 
 
 
 @main_bp.route('/api/updateConductor/<int:id>', methods=['PUT'])
 @verificar_token
-def actualizar_conductor(id):
-    from models import Conductor
-    conductor = Conductor.query.get(id)
-    if conductor is None:
-        return jsonify({'error': 'Conductor no encontrado'}), 404
+def actualizar_conductor(id, id_rol):
 
-    # Obtener los datos del formulario
-    nombre = request.form.get('nombre')
-    patente = request.form.get('patente')
-    auto = request.form.get('auto')
-    nueva_foto = request.files.get('foto')
+    if(id_rol == 2):
+        from models import Conductor
+        conductor = Conductor.query.get(id)
+        if conductor is None:
+            return jsonify({'error': 'Conductor no encontrado'}), 404
 
-    # Actualizar la foto si se proporciona una nueva
-    if nueva_foto:
-        conductor.save_foto(nueva_foto)
+        # Obtener los datos del formulario
+        nombre = request.form.get('nombre')
+        patente = request.form.get('patente')
+        auto = request.form.get('auto')
+        nueva_foto = request.files.get('foto')
 
-    # Actualizar los otros campos si se proporcionan
-    if nombre:
-        conductor.nombre_conductor = nombre
-    if patente:
-        conductor.patente = patente
-    if auto:
-        conductor.nombre_vehiculo = auto
+        # Actualizar la foto si se proporciona una nueva
+        if nueva_foto:
+            conductor.save_foto(nueva_foto)
 
-    db.session.commit()
+        # Actualizar los otros campos si se proporcionan
+        if nombre:
+            conductor.nombre_conductor = nombre
+        if patente:
+            conductor.patente = patente
+        if auto:
+            conductor.nombre_vehiculo = auto
 
-    return jsonify({'message': 'Conductor actualizado correctamente'}), 200
+        db.session.commit()
+
+        return jsonify({'message': 'Conductor actualizado correctamente'}), 200
+    else:
+        
+        return jsonify({'message': 'No tienes el permiso para actualizar el conductor'}), 400
 
 
 @main_bp.route('/api/obtenerConductor/<string:patente>', methods=['GET'])
@@ -105,13 +116,17 @@ def obtenerConductor(patente):
 
 @main_bp.route('/api/borrarConductor/<int:id>', methods=['DELETE'])
 @verificar_token
-def borrarConductor(id):
-    from models import Conductor
-    conductor = Conductor.query.filter_by(id=id).first()
+def borrarConductor(id, id_rol):
+    if(id_rol == 2):
 
-    if conductor:
-        db.session.delete(conductor)
-        db.session.commit()
-        return jsonify({"message": "Conductor borrado exitosamente"})
+        from models import Conductor
+        conductor = Conductor.query.filter_by(id=id).first()
+
+        if conductor:
+            db.session.delete(conductor)
+            db.session.commit()
+            return jsonify({"message": "Conductor borrado exitosamente"})
+        else:
+            return jsonify({"message": "Conductor no encontrado"}), 404
     else:
-        return jsonify({"message": "Conductor no encontrado"}), 404
+       return jsonify({"message": "No tienes el permiso para actualizar un conductor"}), 400
